исправил замечания
