+++ packages/client/src/components/Canvas/Canvas.tsx
@@ -1,34 +1,25 @@
 import React, { useRef, useEffect } from 'react'
-
-interface Mouse {
-  x: number
-  y: number
-}
-interface EnemyTypes {
-  x: number
-  y: number
-  height: number
-  width: number
-  waypointIndex: number
-  center: { x: number; y: number }
-  radius: number
-}
-
-type TowerPlaceType = {
-  x: number
-  y: number
-  size: number
-  color: string
-  occupied: boolean
-}
-type BuildPlace = {
-  x: number
-  y: number
-  isOccupied: boolean
-}
+import { Rect, Enemy, TowerPlace, BuildTower } from './classes'
+import { BuildPlace, TowerPlaceType } from './interfaces'
+import {
+  height,
+  width,
+  offset,
+  enemies,
+  waypoints,
+  towerPlace,
+  towersPlace,
+  buildingTower,
+  mouse,
+} from './consts'
 
 const Canvas: React.FC = () => {
   const refCanvas = useRef<HTMLCanvasElement | null>(null)
+
+  let activeBuildPlace: BuildPlace | TowerPlaceType | null
+
+  let animationFrame = 0
+  let stopGame = false
 
   useEffect(() => {
     const canvas = refCanvas.current
@@ -40,279 +31,8 @@ const Canvas: React.FC = () => {
       return
     }
 
-    const waypoints = [
-      { x: 0, y: 210 },
-      { x: 800, y: 210 },
-    ]
-    const towerPlace = [
-      { x: 250, y: 100 },
-      { x: 350, y: 100 },
-      { x: 450, y: 100 },
-    ]
-
-    const height = 450
-    const width = 800
-
     canvas.height = height
     canvas.width = width
-
-    const offset = 50
-
-    class Rect {
-      x: number
-      y: number
-      color: string
-      w: number
-      h: number
-      active: boolean
-      center: { x: number; y: number }
-
-      constructor(x: number, y: number, w: number, h: number, color: string) {
-        this.x = x
-        this.y = y
-        this.color = color
-
-        this.w = w
-        this.h = h
-
-        this.active = true
-
-        this.center = {
-          x: this.x + this.w / 2,
-          y: this.y + this.h / 2,
-        }
-      }
-
-      draw(context: CanvasRenderingContext2D) {
-        const localContext = context
-
-        localContext.beginPath()
-
-        localContext.fillStyle = this.color
-        localContext.fillRect(this.x, this.y, this.w, this.h)
-
-        localContext.fillStyle = 'rgba(1,1,1,0.5)'
-        localContext.strokeRect(this.x, this.y, this.w, this.h)
-
-        localContext.closePath()
-      }
-
-      update() {
-        if (context) {
-          this.draw(context)
-        }
-      }
-    }
-
-    class Enemy {
-      x: number
-      y: number
-      height: number
-      width: number
-      waypointIndex: number
-      center: { x: number; y: number }
-      radius: number
-
-      constructor(x: number, y: number) {
-        this.x = x
-        this.y = y
-        this.height = 30
-        this.width = 30
-        this.waypointIndex = 0
-        this.center = {
-          x: this.x + this.width / 2,
-          y: this.y + this.height / 2,
-        }
-        this.radius = 15
-      }
-
-      draw(context: CanvasRenderingContext2D) {
-        if (context) {
-          const localContext = context
-
-          localContext.fillStyle = 'red'
-          localContext.beginPath()
-          localContext.arc(
-            this.center.x,
-            this.center.y,
-            this.radius,
-            0,
-            Math.PI * 2
-          )
-          localContext.fill()
-        }
-      }
-
-      update() {
-        if (context) {
-          this.draw(context)
-
-          const waypoint = waypoints[this.waypointIndex]
-          const yDistance = waypoint.y - this.y
-          const xDistance = waypoint.x - this.x
-          const angle = Math.atan2(yDistance, xDistance)
-
-          this.x += Math.cos(angle)
-          this.y += Math.sin(angle)
-
-          if (
-            Math.round(this.x) === Math.round(waypoint.x) &&
-            Math.round(this.y) === Math.round(waypoint.y) &&
-            this.waypointIndex !== waypoints.length
-          ) {
-            this.waypointIndex++
-          }
-
-          //надо будет переделать. не правильно выходит из игры враг
-          if (this.waypointIndex === waypoints.length) {
-            this.waypointIndex = 0
-            this.x = 0
-          }
-
-          this.center = {
-            x: this.x + this.width / 2,
-            y: this.y + this.height / 2,
-          }
-        }
-      }
-    }
-
-    class TowerPlace {
-      x: number
-      y: number
-      size: number
-      color: string
-      occupied: boolean
-
-      constructor(x: number, y: number) {
-        this.x = x
-        this.y = y
-        this.size = 50
-        this.color = 'rgba(255,255,255,0.6)'
-        this.occupied = false
-      }
-
-      draw() {
-        if (context) {
-          context.fillStyle = this.color
-          context.fillRect(this.x, this.y, this.size, this.size)
-        }
-      }
-
-      update(mouse: Mouse) {
-        this.draw()
-        if (
-          mouse.x > this.x &&
-          mouse.x < this.x + this.size &&
-          mouse.y > this.y &&
-          mouse.y < this.y + this.size
-        ) {
-          this.color = 'white'
-        } else {
-          this.color = 'rgba(255,255,255,0.6)'
-        }
-      }
-    }
-
-    class Shot {
-      x: number
-      y: number
-      enemy: EnemyTypes
-      velocity: {
-        x: number
-        y: number
-      }
-      radius: number
-
-      constructor(x: number, y: number, enemy: EnemyTypes) {
-        this.x = x
-        this.y = y
-        this.velocity = {
-          x: 0,
-          y: 0,
-        }
-        this.enemy = enemy
-        this.radius = 5
-      }
-
-      draw() {
-        if (context) {
-          context.beginPath()
-          context.arc(this.x, this.y, this.radius, 0, Math.PI * 2)
-          context.fillStyle = 'rgb(255,140,0)'
-          context.fill()
-        }
-      }
-
-      update() {
-        this.draw()
-
-        const angle = Math.atan2(
-          this.enemy.center.y - this.y,
-          this.enemy.center.x - this.x
-        )
-
-        const power = 3
-        this.velocity.x = Math.cos(angle) * power
-        this.velocity.y = Math.sin(angle) * power
-
-        this.x += this.velocity.x
-        this.y += this.velocity.y
-      }
-    }
-
-    class BuildTower {
-      x: number
-      y: number
-      center: {
-        x: number
-        y: number
-      }
-      shots: Shot[]
-      radius: number
-      target: EnemyTypes | null = null
-      frames: number
-
-      constructor(x: number, y: number) {
-        this.x = x
-        this.y = y
-        this.center = {
-          x: this.x + offset / 2,
-          y: this.y + offset / 2,
-        }
-        this.shots = []
-        this.radius = 200
-        this.target
-        this.frames = 0
-      }
-      draw() {
-        if (context) {
-          context.fillStyle = 'rgb(0,56,176)'
-          context.fillRect(this.x, this.y, offset, offset)
-
-          context.beginPath()
-          context.arc(
-            this.x + offset / 2,
-            this.y + offset / 2,
-            this.radius,
-            0,
-            Math.PI * 2
-          )
-          context.fillStyle = 'rgba(0,56,176, .5)'
-          context.fill()
-        }
-      }
-      update() {
-        if (context) {
-          this.draw()
-          if (this.frames % 100 === 0 && this.target) {
-            this.shots.push(new Shot(this.center.x, this.center.y, this.target))
-          }
-
-          this.frames++
-        }
-      }
-    }
 
     function createGameArea() {
       for (let i = 0; i < width / offset; i++) {
@@ -322,7 +42,8 @@ const Canvas: React.FC = () => {
             t * offset,
             offset,
             offset,
-            'rgba(24,83,98,0.4)'
+            'rgba(24,83,98,0.4)',
+            context
           )
           if (context) {
             rect.draw(context)
@@ -333,29 +54,37 @@ const Canvas: React.FC = () => {
 
     function createGamePath() {
       for (let i = 0; i < width / offset; i++) {
-        const rect = new Rect(i * offset, 200, offset, offset, 'rgb(26,243,68)')
+        const rect = new Rect(
+          i * offset,
+          200,
+          offset,
+          offset,
+          'rgb(26,243,68)',
+          context
+        )
         if (context) {
           rect.draw(context)
         }
       }
     }
 
-    const enemies: Enemy[] = []
-    const towersPlace: TowerPlace[] = []
-    const buildingTower: BuildTower[] = []
-
-    let activeBuildPlace: BuildPlace | TowerPlaceType | null
-
     for (let i = 0; i < 5; i++) {
       const offset = 150
-      enemies.push(new Enemy(waypoints[0].x - offset * i, waypoints[0].y))
+      enemies.push(
+        new Enemy(waypoints[0].x - offset * i, waypoints[0].y, context)
+      )
     }
 
     for (let i = 0; i < towerPlace.length; i++) {
-      towersPlace.push(new TowerPlace(towerPlace[i].x, towerPlace[i].y))
+      towersPlace.push(
+        new TowerPlace(towerPlace[i].x, towerPlace[i].y, context)
+      )
     }
 
     const updateAnimation = function () {
+      if (stopGame) {
+        return
+      }
       context.clearRect(0, 0, width, height)
 
       createGameArea()
@@ -363,6 +92,11 @@ const Canvas: React.FC = () => {
 
       enemies.forEach(enemy => {
         enemy.update()
+
+        if (enemy.x === width) {
+          // пока это условный конец игры когда враг доходит из точки А в точнку Б
+          stopGame = true
+        }
       })
 
       towersPlace.forEach(tower => {
@@ -371,8 +105,6 @@ const Canvas: React.FC = () => {
 
       buildingTower.forEach(building => {
         building.update()
-        // eslint-disable-next-line
-        building.target = null
 
         const validEnemies = enemies.filter(enemy => {
           const xDiff = enemy.center.x - building.center.x
@@ -397,24 +129,19 @@ const Canvas: React.FC = () => {
         }
       })
 
-      requestAnimationFrame(updateAnimation)
+      animationFrame = requestAnimationFrame(updateAnimation)
     }
 
-    const mouse: Mouse = {
-      x: 0,
-      y: 0,
-    }
-
-    canvas.addEventListener('click', () => {
+    const buildTowerHandler = () => {
       if (activeBuildPlace && !(activeBuildPlace as BuildPlace).isOccupied) {
         buildingTower.push(
-          new BuildTower(activeBuildPlace.x, activeBuildPlace.y)
+          new BuildTower(activeBuildPlace.x, activeBuildPlace.y, context)
         )
         ;(activeBuildPlace as BuildPlace).isOccupied = true
       }
-    })
+    }
 
-    window.addEventListener('mousemove', (e: MouseEvent) => {
+    const mouseActiveBuildPlace = (e: MouseEvent) => {
       mouse.x = e.clientX
       mouse.y = e.clientY
 
@@ -432,9 +159,19 @@ const Canvas: React.FC = () => {
           break
         }
       }
-    })
+    }
+
+    canvas.addEventListener('click', buildTowerHandler)
+    window.addEventListener('mousemove', mouseActiveBuildPlace)
 
     updateAnimation()
+
+    return () => {
+      canvas.removeEventListener('click', buildTowerHandler)
+      window.removeEventListener('mousemove', mouseActiveBuildPlace)
+
+      cancelAnimationFrame(animationFrame)
+    }
   }, [])
 
   return <canvas width={800} height={450} ref={refCanvas} />
